+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/pdom/db/IString.java
@@ -25,7 +25,7 @@ public interface IString {
 	/**
 	 * Get the offset of this IString record in the PDOM
 	 */
-	public int getRecord();
+	public long getRecord();
 	
 	// strcmp equivalents
 	/**
