+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/pdom/dom/MacroContainerCollector.java
@@ -53,7 +53,7 @@ public final class MacroContainerCollector implements IBTreeVisitor {
 		monitor= pm;
 	}
 	
-	final public int compare(int record) throws CoreException {
+	final public int compare(long record) throws CoreException {
 		if (monitor != null)
 			checkCancelled();
 		IString rhsName= PDOMNamedNode.getDBName(linkage.getDB(), record);
@@ -79,7 +79,7 @@ public final class MacroContainerCollector implements IBTreeVisitor {
 		return cmp;
 	}
 	
-	final public boolean visit(int record) throws CoreException {
+	final public boolean visit(long record) throws CoreException {
 		if (monitor != null)
 			checkCancelled();
