+++ core/org.eclipse.cdt.core/parser/org/eclipse/cdt/internal/core/pdom/dom/cpp/PDOMCPPConstructor.java
@@ -25,7 +25,7 @@ class PDOMCPPConstructor extends PDOMCPPMethod implements ICPPConstructor {
 		super(linkage, parent, method);
 	}
 
-	public PDOMCPPConstructor(PDOMLinkage linkage, int record) {
+	public PDOMCPPConstructor(PDOMLinkage linkage, long record) {
 		super(linkage, record);
 	}
